b2b_portal/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.2
mysqlclient>=2.2
celery>=5.4
redis>=5.0
python-dotenv>=1.0
django-cors-headers>=4.4
gunicorn>=21.2
django-jazzmin>=2.6
whitenoise>=6.7
